feat: fix sidebar toggle icons and enhance task scheduling, labelling, scroll bar issue when side bar is toggled off and filtering of tasks.
