--- StudyPlan-contribution-01-main/.env.example
@@ -1,8 +1,0 @@
-# StudyPlan Environment Configuration
-
-# Google Gemini API Key
-# Get your key at: https://aistudio.google.com/app/apikey
-GEMINI_API_KEY=your_gen_ai_key_here
-
-# Server Port (optional, defaults to 3000)
-PORT=3000
