--- StudyPlan-contribution-01-main/.github/ISSUE_TEMPLATE/bug_report.md
@@ -1,33 +1,0 @@
----
-name: Bug Report
-about: Create a report to help us improve
-title: 'bug: '
-labels: bug
-assignees: ''
-
----
-
-# Problem
-Describe the bug clearly.
-
-# Current Behavior
-Explain what is happening currently.
-
-# Expected Behavior
-Explain what should happen instead.
-
-# Steps to Reproduce
-1. Go to '...'
-2. Click on '....'
-3. Scroll down to '....'
-4. See error
-
-# Screenshots
-If applicable, add screenshots to help explain your problem.
-
-# Environment
-- Browser: [e.g. Chrome, Safari]
-- Version: [e.g. 22]
-
-# Additional Context
-Add any other context about the problem here.
