--- StudyPlan-contribution-01-main/.github/ISSUE_TEMPLATE/feature_request.md
@@ -1,26 +1,0 @@
----
-name: Feature Request
-about: Suggest an idea for this project
-title: 'feat: '
-labels: enhancement
-assignees: ''
-
----
-
-# Problem
-Describe the problem or limitation you are facing.
-
-# Proposed Improvement
-Explain the planned fix/improvement.
-
-# Why improvement is needed
-Explain the benefit of this feature.
-
-# Expected Result
-Describe the expected outcome after implementation.
-
-# Alternatives Considered
-A clear and concise description of any alternative solutions or features you've considered.
-
-# Additional Context
-Add any other context or screenshots about the feature request here.
