--- StudyPlan-contribution-01-main/.github/pull_request_template.md
@@ -1,21 +1,0 @@
-# Related Issue
-Closes #ISSUE_NUMBER
-
-# Summary
-Brief explanation of the contribution.
-
-# Changes Made
-- Bullet list of implemented changes.
-- ...
-
-# Testing
-Explain how the changes were tested.
-
-# Screenshots
-Add screenshots if UI changes exist.
-
-# Checklist
-- [x] Code follows project style
-- [x] Tested locally
-- [x] No unrelated changes included
-- [x] Documentation updated (if applicable)
