--- StudyPlan-contribution-01-main/.gitignore
@@ -1,22 +1,0 @@
-# Node.js
-node_modules/
-npm-debug.log
-yarn-error.log
-package-lock.json
-
-# Environment Variables
-.env
-.env.local
-
-# Database
-studyplan.db
-studyplan.db-journal
-
-# OS generated files
-.DS_Store
-Thumbs.db
-
-# Optional Editor configurations
-.vscode/
-.idea/
-studyplan.db
