--- StudyPlan-contribution-01-main/.render-build.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-npm install
-npm rebuild sqlite3 --build-from-source
