--- StudyPlan-contribution-01-main/backend/routers/csvDownload.router.js
@@ -1,12 +1,0 @@
-const express = require('express');
-const {
-  downloadData,
-  downloadCalendar,
-} = require('../controllers/csvDownload.controller.js');
-
-const router = express.Router();
-
-router.get('/download', downloadData);
-router.get('/download/calendar', downloadCalendar);
-
-module.exports = router;
