--- StudyPlan-contribution-01-main/error.html
@@ -1,98 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Something Went Wrong | StudyPlan</title>
-  <style>
-    :root {
-      --bg: #f4f2ec;
-      --card: #ffffff;
-      --text: #1f1b17;
-      --muted: #655f57;
-      --accent: #2d5da0;
-      --border: rgba(0, 0, 0, 0.11);
-    }
-
-    * { box-sizing: border-box; }
-
-    body {
-      margin: 0;
-      min-height: 100vh;
-      display: grid;
-      place-items: center;
-      padding: 24px;
-      font-family: "Inter", system-ui, -apple-system, sans-serif;
-      color: var(--text);
-      background:
-        radial-gradient(circle at 85% 8%, rgba(45, 93, 160, 0.16), transparent 35%),
-        radial-gradient(circle at 12% 88%, rgba(133, 79, 11, 0.12), transparent 38%),
-        var(--bg);
-    }
-
-    .card {
-      width: min(620px, 100%);
-      background: var(--card);
-      border: 1px solid var(--border);
-      border-radius: 18px;
-      padding: 28px;
-      box-shadow: 0 16px 38px rgba(0, 0, 0, 0.08);
-    }
-
-    h1 {
-      font-size: 32px;
-      margin: 0 0 12px;
-    }
-
-    p {
-      margin: 0 0 12px;
-      color: var(--muted);
-      line-height: 1.6;
-    }
-
-    .quote {
-      color: var(--accent);
-      font-weight: 700;
-      margin-top: 4px;
-    }
-
-    .actions {
-      display: flex;
-      gap: 12px;
-      margin-top: 14px;
-      flex-wrap: wrap;
-    }
-
-    button, a {
-      border: none;
-      border-radius: 10px;
-      padding: 10px 14px;
-      font-size: 14px;
-      font-weight: 700;
-      text-decoration: none;
-      cursor: pointer;
-    }
-
-    button {
-      color: #fff;
-      background: var(--accent);
-    }
-
-    a {
-      color: var(--text);
-      background: #ece9e2;
-    }
-  </style>
-</head>
-<body>
-  <main class="card">
-    <h1>Something went wrong</h1>
-    <p>The app ran into an unexpected problem while loading this page.</p>
-    <p class="quote">We&apos;ll be back, keep studying, keep learning.</p>
-    <div class="actions">
-      <button type="button" onclick="window.location.reload()">Refresh Page</button>
-      <a href="/">Back to Dashboard</a>
-    </div>
-  </main>
-</body>
-</html>
