--- StudyPlan-contribution-01-main/js/utils/api.js
@@ -1,19 +1,0 @@
-export async function extractTasksFromText(text) {
-  try {
-    const res = await fetch('/api/extract', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ text })
-    });
-    
-    if (!res.ok) {
-      console.error('Extraction failed', await res.text());
-      return [];
-    }
-    
-    return await res.json();
-  } catch (e) {
-    console.error('Error hitting extract endpoint', e);
-    return [];
-  }
-}
