--- StudyPlan-contribution-01-main/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "study-plan",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "node server.js",
-    "test": "node --test"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "@google/genai": "^1.49.0",
-    "cors": "^2.8.6",
-    "dotenv": "^17.4.2",
-    "express": "^5.2.1",
-    "sqlite3": "^6.0.1"
-  },
-  "engines": {
-    "node": ">=20.x"
-  }
-}
