--- StudyPlan-contribution-01-main/support-page/support.js
@@ -1,1 +1,0 @@
-console.log("Support page loaded");
