--- StudyPlan-contribution-01-main/support-page/terms.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-  <title>Terms & Conditions | StudyPlan</title>
-
-  <link rel="stylesheet" href="/support-page/support.css">
-
-  <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap" rel="stylesheet">
-</head>
-
-<body>
-
-<header class="site-header">
-
-  <div class="header-left">
-    <img src="/logo.png" alt="Logo" class="logo">
-    <h1 class="site-title">StudyPlan</h1>
-  </div>
-
-  <nav class="header-nav">
-    <a href="/" class="nav-btn">Dashboard</a>
-  </nav>
-
-</header>
-
-<main class="support-wrapper">
-
-  <section class="card">
-
-    <h2>Terms & Conditions</h2>
-
-    <p>
-      By using StudyPlan, you agree to use the service responsibly.
-      Users are responsible for verifying deadlines extracted by AI before submission.
-    </p>
-
-    <p>
-      StudyPlan is not liable for missed deadlines caused by incorrect user input or AI inference errors.
-    </p>
-
-  </section>
-
-</main>
-
-<footer class="site-footer">
-
-  <p>© 2026 StudyPlan. All rights reserved.</p>
-
-  <div class="footer-links">
-    <a href="/support-page/privacy.html">Privacy</a>
-    <a href="/support-page/terms.html">Terms</a>
-    <a href="/support-page/support.html">Support</a>
-  </div>
-
-</footer>
-
-<script src="/support-page/support.js"></script>
-
-</body>
-</html>
